+++ tests/testthat/test-ptransform.R
@@ -1,0 +1,7 @@
+context("ptransform")
+
+library(maptools)
+data(wrld_simpl)
+test_that("transforming works", {
+  expect_equal(2 * 2, 4)
+})